Add files via upload
Run this in a java compiler to make it work!

Make sure that you have a text file named deck.txt!
Will make an exe of this but for now this is just a file that will run in a java compiler
ALSO make sure you leave out the extra deck and side deck! ONLY KEEP THE MAIN DECK!

You can change the value from 1000 to any amount.